remove player movement + bg img. create ani plugin
