+++ src/game/animations/components.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::*;
+
+// All animated sprites must use this
+#[derive(Component)]
+pub struct AnimateSprite;
