+++ src/game/player/mod.rs
@@ -14,8 +14,6 @@ impl Plugin for PlayerPlugin {
         app
             // On enter InGame state
             .add_systems(OnEnter(AppState::InGame), spawn_player)
-            // Systems
-            .add_systems(Update, player_movement.run_if(in_state(AppState::InGame)))
             // On exit InGame state
             .add_systems(OnExit(AppState::InGame), despawn_player);
     }
